--- a/recipes/lora_finetune_single_device.py
+++ b/recipes/lora_finetune_single_device.py
@@ -18,7 +18,6 @@ from omegaconf import DictConfig, ListConfig
 from torch import nn
 from torch.optim import Optimizer
 from torch.utils.data import DataLoader, DistributedSampler
-from torchao.dtypes.nf4tensor import NF4Tensor
 from torchtune import config, modules, training, utils
 from torchtune.data import padded_collate_sft
 from torchtune.datasets import ConcatDataset
@@ -287,22 +286,29 @@ class LoRAFinetuneRecipeSingleDevice(FTRecipeInterface):
         # Set up profiler, returns DummyProfiler (nullcontext object with no-op `step` method)
         # if cfg is missing profiler key or if `cfg.profiler.enabled = False
         self._profiler = self._setup_profiler(cfg.get(PROFILER_KEY, None))
-    
+
     def _patch_state_dict_hook(self, cfg: DictConfig) -> None:
         """
         Patches the state_dict hook for QLoRA for the `low_cpu_ram` config option.
         """
         self._previous_hook = None
         if cfg.get("low_cpu_ram", False):
-            if common_utils._low_ram_reparametrize_as_dtype_state_dict_post_hook is None:
+            if (
+                common_utils._low_ram_reparametrize_as_dtype_state_dict_post_hook
+                is None
+            ):
                 if sys.platform == "win32":
-                    raise RuntimeError(
-                        "low_cpu_ram=True not supported on Windows."
-                    )
+                    raise RuntimeError("low_cpu_ram=True not supported on Windows.")
                 else:
-                    raise RuntimeError("low_cpu_ram=True requires torch.__version__ >= 2.5.0.dev20240830.")
-            self._previous_hook = common_utils.reparametrize_as_dtype_state_dict_post_hook
-            common_utils.reparametrize_as_dtype_state_dict_post_hook = common_utils._low_ram_reparametrize_as_dtype_state_dict_post_hook
+                    raise RuntimeError(
+                        "low_cpu_ram=True requires torch.__version__ >= 2.5.0.dev20240830."
+                    )
+            self._previous_hook = (
+                common_utils.reparametrize_as_dtype_state_dict_post_hook
+            )
+            common_utils.reparametrize_as_dtype_state_dict_post_hook = (
+                common_utils._low_ram_reparametrize_as_dtype_state_dict_post_hook
+            )
 
         # Used to ignore labels for loss computation
         self.ignore_labels_cache = torch.full(
@@ -730,7 +736,9 @@ class LoRAFinetuneRecipeSingleDevice(FTRecipeInterface):
 
     def cleanup(self) -> None:
         if self._previous_hook is not None:
-            common_utils.reparametrize_as_dtype_state_dict_post_hook = self._previous_hook
+            common_utils.reparametrize_as_dtype_state_dict_post_hook = (
+                self._previous_hook
+            )
             self._previous_hook = None
         self._metric_logger.close()