#### Context
This PR prevents OOM during checkpoint save on colab for the following recipe, 

`tune run recipes/lora_finetune_single_device.py --config recipes/configs/llama3/8B_qlora_single_device.yaml`

I do believe it is possible to do better (perf improvements for this colab case) if we refactor the checkpointing logic, but that would be a more invasive refactor imo for now this is the most minimally invasive change that unblocks this use case.


#### Changelog
- Added a new hook `_low_ram_reparametrize_as_dtype_state_dict_post_hook` and `_register_reparametrize_state_dict_hooks` that toggles between the regular reparametrize state_dict hook and this one
- added a model config `low_cpu_ram` -- when `low_cpu_ram` is`True` `_register_reparametrize_state_dict_hooks` toggles to the `low_ram` version of the hook
- PyTorch side changes were landed in https://github.com/pytorch/pytorch/pull/134504 [was reverted, remerging right now as of 9/05] and https://github.com/pytorch/pytorch/pull/134371 **both should be available in the 20240830 nightly**

Old changelog (keeping around for posterity)
- ~Set a seed  in `8B_qlora_single_device.yaml` to make dataloader samples (and hence weights) deterministic~
- ~[for testing velocity purposes]Changed number of epochs to 8, reduced `max_steps_per_epoch` --> 20, `gradient_accumulation_steps` --> 2 to make `save_checkpoint` be called sooner~
- ~[for loss curves] some changes to lora_finetune_single_device.py to mimic a user doing `resume_from_checkpoint` after each epoch (while preserving logger)~
- ~[Not to be landed in torchtune, for PoC purpose] patched [`FakeTensor.__reduce_ex__ `](https://github.com/pytorch/torchtune/pull/1315/files#diff-734dbb110089223adfc465adb1594c2645bc025171e23c67f7b7bece857edbf0R42-R56) which is needed to ensure the `write_record_metadata` utility to create empty checkpoints is called (requires changes in https://github.com/pytorch/pytorch/pull/133272) I need to figure out how to land this piece :)~ 
- ~Skip registration of `reparametrize_as_dtype_state_dict_post_hook` for llama3~
~- Showed an example of how to do the corresponding (using mmap to prevent OOM) in `lora_finetune_single_device.py:save_checkpoint`~

#### Test plan

##### Sanity check 
Ran `tune run` command on devgpu  (with only the changes in `8B_qlora_single_device.yaml` to set seed and decrease steps per epoch)  and verified that `meta_model_0.pt` generated is the same before and after the changes in this PR with small [snippet](https://gist.github.com/mikaylagawarecki/60cf8e77f6230fcd45c62a4a7f479370)

| | Before ckpt save time (s) | After ckpt save time (s) |
|-|-|-|
| devgpu | 178s | ~200s (ranged from 190 to 220ish) |
|colab | OOM (no baseline) | 510s (ranged from 421s to 610s) |

Verified that colab does not OOM
https://colab.research.google.com/drive/1y7Az78ATauK7gkewZkcMO3cNgVWm1233?usp=sharing


##### Loss Curves
The validation was run on commit `abdbd7` which has special logic to mimic `resume_from_checkpoint` for each epoch

Config is per the changes in `recipes/configs/llama3/8B_qlora_single_device.yaml` (8 epochs with 20 steps per epoch and gradient accumulation every 2 steps) on 6cf31b6. For the reloading checkpoint case [I modified `lora_fine_tune_single_device.py:recipe_main` to mimic `resume_from_checkpoint` after each epoch](https://github.com/pytorch/torchtune/pull/1315/files#diff-734dbb110089223adfc465adb1594c2645bc025171e23c67f7b7bece857edbf0R711-R727)

Devgpu:
<img width="1456" alt="Screenshot 2024-09-05 at 10 29 09 PM" src="https://github.com/user-attachments/assets/e6fc5533-e3cc-4dae-b38a-9a69da94c935">



Colab:
<img width="1451" alt="Screenshot 2024-09-05 at 10 26 19 PM" src="https://github.com/user-attachments/assets/c8ac54aa-afff-4522-93dd-2e0536aa8b7f">