@mikaylagawarecki 2024-09-10
Added changes from https://github.com/pytorch/torchtune/pull/1535 + one more round of loss curves

<img width="434" alt="Screenshot 2024-09-10 at 3 27 23 PM" src="https://github.com/user-attachments/assets/fb131177-d2de-427a-9e75-c7dae0184c1d">
<img width="433" alt="Screenshot 2024-09-10 at 3 27 12 PM" src="https://github.com/user-attachments/assets/4b9d3947-d9c3-48ab-995c-9010f639ebc4">